change type of returnval to long, add false testcases

+++ c/http_server/tests/includes/test_source_http.h
@@ -5,7 +5,7 @@
 
 typedef struct{
 	http_header parsed_header;
-	int returnval;
+	long returnval;
 }result_parse_next_http_header;
 typedef struct{
 	char *buffer;
